2nd set of review comments handled
